requirements.txt
anyio==3.7.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.99.0
future==0.18.3
h11==0.14.0
httptools==0.5.0
idna==3.4
iso8601==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==1.10.10
pyserial==3.5
python-dotenv==1.0.0
PyYAML==6.0
serial==0.0.97
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
